fix cmake compile Event lib error; rename Event to EventSystem
