--- Engine/Core/Event/CMakeLists.txt
@@ -1,3 +1,0 @@
-# created on 2021/7/27
-# author @zoloypzuo
-add_library(Event Event.h Event.inl)
