+++ Engine/Core/EventSystem/CMakeLists.txt
@@ -1,0 +1,3 @@
+# created on 2021/7/27
+# author @zoloypzuo
+add_library(EventSystem Event.cpp Event.h Event.inl Event.cpp)
